backend_py/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
openai==1.43.0

# For Google Gemini (pick one, new preferred)
google-genai>=0.3.0; python_version>='3.8'
google-generativeai>=0.7.2; python_version>='3.8'

# Image processing
Pillow>=10.0.0
pillow-avif-plugin>=1.4.7
pillow-heif>=0.16.0

# Data processing
numpy>=1.26.0

# Database
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9
asyncpg==0.29.0  # PostgreSQL 비동기 드라이버

# Authentication
google-auth>=2.3.0

# Redis caching (NEW)
redis>=5.0.0  # Redis 캐싱 시스템용
